requirements.txt
flask
crawl4ai
google
python-dotenv
